pyproject.toml
[project]
name = "dentistry"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.43",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
    "beautifulsoup4>=4.12.2",
    "python-dotenv>=1.0.0",
]

